🚚 F04: ORDER > Move Navbar to navbar folder
