+++ src/components/pages/OrderPage.jsx
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import Navbar from "../Navbar"
+import Navbar from "./navbar/Navbar"
 import { theme } from "../../theme";
 
 const OrderPage = () => {
